+++ src/main/java/org/jungppo/bambooforest/chat/domain/entity/ChatRoomEntity.java
@@ -1,6 +1,4 @@
-package org.jungppo.bambooforest.entity.chat;
-
-import org.jungppo.bambooforest.entity.common.JpaBaseEntity;
+package org.jungppo.bambooforest.chat.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,15 +7,16 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Getter;
+import org.jungppo.bambooforest.entity.common.JpaBaseEntity;
 
 @Entity
 @Getter
 @Table(name = "chat_room")
 public class ChatRoomEntity extends JpaBaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
-	@Column(nullable = false)
-	private String name;
+    @Column(nullable = false)
+    private String name;
 }
