--- src/main/java/org/jungppo/bambooforest/entity/chat/ChatMessageEntity.java
@@ -1,35 +1,0 @@
-package org.jungppo.bambooforest.entity.chat;
-
-import org.jungppo.bambooforest.entity.common.JpaBaseEntity;
-import org.jungppo.bambooforest.member.domain.MemberEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-
-@Entity
-@Getter
-@Table(name = "chat_message")
-public class ChatMessageEntity extends JpaBaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY, optional = false)
-	@JoinColumn(name = "chat_room_id", nullable = false)
-	private ChatRoomEntity chatRoom;
-
-	@ManyToOne(fetch = FetchType.LAZY, optional = false)
-	@JoinColumn(name = "member_id", nullable = false)
-	private MemberEntity member;
-
-	@Column(nullable = false)
-	private String content;
-}
