--- src/main/java/org/jungppo/bambooforest/member/domain/RoleType.java
@@ -1,5 +1,0 @@
-package org.jungppo.bambooforest.member.domain;
-
-public enum RoleType {
-    ROLE_USER, ROLE_ADMIN
-}
