+++ src/main/java/org/jungppo/bambooforest/member/domain/entity/OAuth2Type.java
@@ -1,4 +1,4 @@
-package org.jungppo.bambooforest.member.domain;
+package org.jungppo.bambooforest.member.domain.entity;
 
 import java.util.Arrays;
 import java.util.Collections;
