+++ src/main/java/org/jungppo/bambooforest/member/domain/entity/RoleType.java
@@ -1,0 +1,5 @@
+package org.jungppo.bambooforest.member.domain.entity;
+
+public enum RoleType {
+    ROLE_USER, ROLE_ADMIN
+}
