+++ src/main/java/org/jungppo/bambooforest/member/domain/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
-package org.jungppo.bambooforest.repository.member;
+package org.jungppo.bambooforest.member.domain.repository;
 
-import org.jungppo.bambooforest.member.domain.RefreshTokenEntity;
+import org.jungppo.bambooforest.member.domain.entity.RefreshTokenEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
