+++ src/main/java/org/jungppo/bambooforest/payment/domain/repository/PaymentRepository.java
@@ -1,7 +1,7 @@
-package org.jungppo.bambooforest.repository.payment;
+package org.jungppo.bambooforest.payment.domain.repository;
 
 import java.util.UUID;
-import org.jungppo.bambooforest.payment.domain.PaymentEntity;
+import org.jungppo.bambooforest.payment.domain.entity.PaymentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<PaymentEntity, UUID> {
