+++ src/main/java/org/jungppo/bambooforest/payment/service/PaymentService.java
@@ -6,15 +6,15 @@ import org.jungppo.bambooforest.battery.domain.BatteryItem;
 import org.jungppo.bambooforest.client.paymentgateway.PaymentGatewayClient;
 import org.jungppo.bambooforest.dto.paymentgateway.PaymentResponse;
 import org.jungppo.bambooforest.dto.paymentgateway.toss.TossPaymentRequest;
-import org.jungppo.bambooforest.member.domain.MemberEntity;
+import org.jungppo.bambooforest.member.domain.entity.MemberEntity;
+import org.jungppo.bambooforest.member.domain.repository.MemberRepository;
 import org.jungppo.bambooforest.member.dto.PaymentConfirmRequest;
 import org.jungppo.bambooforest.member.dto.PaymentDto;
 import org.jungppo.bambooforest.member.dto.PaymentSetupRequest;
 import org.jungppo.bambooforest.member.dto.PaymentSetupResponse;
-import org.jungppo.bambooforest.payment.domain.PaymentEntity;
-import org.jungppo.bambooforest.payment.domain.PaymentStatusType;
-import org.jungppo.bambooforest.repository.member.MemberRepository;
-import org.jungppo.bambooforest.repository.payment.PaymentRepository;
+import org.jungppo.bambooforest.payment.domain.entity.PaymentEntity;
+import org.jungppo.bambooforest.payment.domain.entity.PaymentStatusType;
+import org.jungppo.bambooforest.payment.domain.repository.PaymentRepository;
 import org.jungppo.bambooforest.response.exception.battery.BatteryNotFoundException;
 import org.jungppo.bambooforest.response.exception.member.MemberNotFoundException;
 import org.jungppo.bambooforest.response.exception.payment.PaymentFailureException;
