+++ src/main/java/org/jungppo/bambooforest/security/oauth2/CustomOAuth2LoginSuccessHandler.java
@@ -10,8 +10,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import lombok.extern.slf4j.Slf4j;
-import org.jungppo.bambooforest.member.domain.OAuth2Type;
-import org.jungppo.bambooforest.member.domain.RoleType;
+import org.jungppo.bambooforest.member.domain.entity.OAuth2Type;
+import org.jungppo.bambooforest.member.domain.entity.RoleType;
 import org.jungppo.bambooforest.member.service.RefreshTokenService;
 import org.jungppo.bambooforest.security.jwt.JwtMemberClaim;
 import org.jungppo.bambooforest.util.CookieUtils;
