+++ src/main/java/org/jungppo/bambooforest/security/oauth2/response/GitHubResponse.java
@@ -1,12 +1,12 @@
 package org.jungppo.bambooforest.security.oauth2.response;
 
-import static org.jungppo.bambooforest.member.domain.OAuth2Type.OAUTH2_GITHUB;
+import static org.jungppo.bambooforest.member.domain.entity.OAuth2Type.OAUTH2_GITHUB;
 import static org.jungppo.bambooforest.security.oauth2.constants.GitHubConstants.GITHUB_ID;
 import static org.jungppo.bambooforest.security.oauth2.constants.GitHubConstants.GITHUB_NICKNAME;
 import static org.jungppo.bambooforest.security.oauth2.constants.GitHubConstants.GITHUB_PROFILE_IMAGE_URL;
 
 import java.util.Map;
-import org.jungppo.bambooforest.member.domain.OAuth2Type;
+import org.jungppo.bambooforest.member.domain.entity.OAuth2Type;
 
 public class GitHubResponse implements OAuth2Response {
